+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## **Version 0.1.0** (planned 21 November 2021)
+## **Version 0.1.0** (29 November 2021)
 
 - AG-26, Get on track a whole Angotia sdk
 - AG-31, Fix error handling when `Engine API` is turned off